DVL/Wayfinder/requirements.txt
pyserial>=3.4
numpy

